Mostrar mensajes de confirmación y espera

+++ src/modules/daybook/views/EntryView.vue
@@ -51,6 +51,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -104,6 +105,12 @@ export default {
         },
         async saveEntry() {
 
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
             if ( this.entry.id  ) {
                 // Actualizar
                 await this.updateEntry( this.entry )
@@ -112,10 +119,30 @@ export default {
                 const id = await this.createEntry( this.entry )
                 this.$router.push({ name: 'entry', params: { id } })
             }
+
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
+
         },
         async onDeleteEntry() {
-            await this.deleteEntry( this.entry.id )
-            this.$router.push({ name: 'no-entry' })
+
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
+
+            if ( isConfirmed ) {
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry( this.entry.id )
+                this.$router.push({ name: 'no-entry' })
+
+                Swal.fire('Eliminado','','success')
+            }
         }
     },
     created() {
